Organize w3ctrace code into separate files
